refactor(database): cms.handler.ts to reduce duplicates
