+++ modules/database/src/controllers/cms/schema.controller.ts
@@ -8,7 +8,7 @@ import { isNil } from 'lodash';
 import { DatabaseAdapter } from '../../adapters/DatabaseAdapter';
 import { MongooseSchema } from '../../adapters/mongoose-adapter/MongooseSchema';
 import { SequelizeSchema } from '../../adapters/sequelize-adapter/SequelizeSchema';
-import { CmsHandlers } from '../../handlers/cms.handler';
+import { CmsHandlers } from '../../handlers/cms/cms.handler';
 import { ParsedQuery } from '../../interfaces';
 
 type _ConduitSchema = Omit<ConduitSchema, 'schemaOptions'> & {
