+++ modules/database/src/routes/index.ts
@@ -1,4 +1,4 @@
-import { CmsHandlers } from '../handlers/cms.handler';
+import { CmsHandlers } from '../handlers/cms/cms.handler';
 import ConduitGrpcSdk, {
   ConduitRouteOptions,
   ConduitRouteReturnDefinition,
